+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### 2019-05-24 / 1.1.0
+
+*   feat(core): search with exact match
+*   fix(core): do not load config when connection pool is provided
+*   fix(core): error when migrating with numerical version
+
 ### 2018-05-21 / 1.0.0
 
 *   Initial version.